moving source code to new dir
moving source code from /home/vagrant to /vagrant. This allows
code to be modified from host machine which enables people to
use their preferred IDE and also check coverage reports

Did not move the bin directory as there was a problem with
selinux/template module in ansible. So leaving in /home/vagrant
for now. Should not be an issue since the dir is added to PATH
env. variable.

Signed-off-by: Thiago da Silva <thiago@redhat.com>